requirements.txt
fastapi==0.128.0
pydantic==2.12.5
email-validator==2.1.1
uvicorn==0.41.0
sqlalchemy==2.0.36
python-dotenv==1.2.1
psycopg2==2.9.11
python-jose==3.5.0
passlib==1.7.4
python-multipart==0.0.20